+++ carddav.php
@@ -25,6 +25,7 @@
 declare(strict_types=1);
 
 use MStilkerich\CardDavClient\Account;
+use MStilkerich\CardDavClient\{AddressbookCollection,WebDavResource};
 use Psr\Log\LoggerInterface;
 use MStilkerich\CardDavAddressbook4Roundcube\{Addressbook, Config, RoundcubeLogger};
 use MStilkerich\CardDavAddressbook4Roundcube\Db\{Database, AbstractDatabase};
@@ -280,8 +281,7 @@ class carddav extends rcube_plugin
 
                         $logger->info("Adding preset for $username at URL $url");
                         $account = new Account($url, $username, $password);
-                        $discover = $infra->makeDiscoveryService();
-                        $abooks = $discover->discoverAddressbooks($account);
+                        $abooks = $this->determineAddressbooksToAdd($account);
 
                         foreach ($abooks as $abook) {
                             if ($preset['carddav_name_only']) {
@@ -552,8 +552,7 @@ class carddav extends rcube_plugin
                         $new['username'],
                         self::replacePlaceholdersPassword($new['password'])
                     );
-                    $discover = $infra->makeDiscoveryService();
-                    $abooks = $discover->discoverAddressbooks($account);
+                    $abooks = $this->determineAddressbooksToAdd($account);
 
                     if (count($abooks) > 0) {
                         $basename = $new['name'];
@@ -646,6 +645,20 @@ class carddav extends rcube_plugin
         }
 
         return $ret;
+    }
+
+    /**
+     * Compares the path components of two URIs.
+     *
+     * @return bool True if the normalized path components are equal.
+     */
+    private static function compareUrlPaths(string $url1, string $url2): bool
+    {
+        $comp1 = \Sabre\Uri\parse($url1);
+        $comp2 = \Sabre\Uri\parse($url2);
+        $p1 = trim(rtrim($comp1["path"] ?? "", "/"), "/");
+        $p2 = trim(rtrim($comp2["path"] ?? "", "/"), "/");
+        return $p1 === $p2;
     }
 
     /**
@@ -1255,6 +1268,55 @@ class carddav extends rcube_plugin
             $imap_password .= $imap_password;
         }
         return substr($imap_password, 0, 24);
+    }
+
+    /**
+     * Determines the addressbooks to add for a given URI.
+     *
+     * We perform discovery to determine all the user's addressbooks.
+     *
+     * If the given URI might point to an addressbook directly (i.e. it has a non-empty path), we check if it is
+     * contained in the discovered addressbooks. If it is not, we check if it actually points to an addressbook. If it
+     * does, we add ONLY this addressbook, not the discovered ones.
+     *
+     * We need to perform the discovery to determine where the user's own addressbooks live (addressbook home).
+     *
+     * See https://github.com/mstilkerich/rcmcarddav/issues/339 for rationale.
+     *
+     * @param Account $account The account to discover the addressbooks for. The discovery URI is assumed as user input.
+     * @return list<AddressbookCollection> The determined addressbooks, possible empty.
+     */
+    private function determineAddressbooksToAdd(Account $account): array
+    {
+        $infra = Config::inst();
+        $logger = $infra->logger();
+
+        $uri = $account->getDiscoveryUri();
+
+        $discover = $infra->makeDiscoveryService();
+        $abooks = $discover->discoverAddressbooks($account);
+
+        foreach ($abooks as $abook) {
+            // If the discovery URI points to an addressbook that was also discovered, we use the discovered results
+            // We deliberately only compare the path components, as the server part may contain a port (or not) or even
+            // be a different server name after discovery, but the path part should be "unique enough"
+            if (self::compareUrlPaths($uri, $abook->getUri())) {
+                return $abooks;
+            }
+        }
+
+        // If the discovery URI points to an addressbook that is not part of the discovered ones, we only use that
+        // addressbook
+        try {
+            $directAbook = WebDavResource::createInstance($uri, $account);
+            if ($directAbook instanceof AddressbookCollection) {
+                $logger->debug("Only adding non-individual addressbook $uri");
+                return [ $directAbook ];
+            }
+        } catch (\Exception $e) {
+        }
+
+        return $abooks;
     }
 
     /**